Make stage scrollable
