+++ src/styles/EnvVisualizer.css
@@ -2,6 +2,6 @@
   background-color: #2c3e50;
 }
 
-.sa-env-visualizer {
+#stage {
   overflow: auto;
 }
